backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# AI / NLP
mistralai==0.4.2
sentence-transformers==2.7.0
spacy==3.8.4
scikit-learn==1.4.2
numpy==1.26.4

# Async tasks
celery==5.4.0
redis==5.0.4

# Document parsing
python-pptx==1.0.2
pdfplumber==0.11.4
python-docx==1.1.2

# Utils
python-dotenv==1.0.1
httpx==0.27.0
